checkConnection
